+++ src/util.rs
@@ -1,7 +1,8 @@
 use std::fmt::Debug;
 use std::str::FromStr;
 
-use bitcoin::util::bip32::IntoDerivationPath;
+use bitcoin::hashes::{sha256, Hash};
+use bitcoin::util::bip32::{ChildNumber, IntoDerivationPath};
 use bitcoin::{secp256k1, PublicKey};
 use miniscript::descriptor::{DescriptorPublicKey, DescriptorTrait, Wildcard};
 use miniscript::{bitcoin, ForEachKey, TranslatePk2};
@@ -137,6 +138,37 @@ impl DeriveExt for Vec<Value> {
     fn is_deriveable(&self) -> bool {
         self.iter().any(|v| v.is_deriveable())
     }
+}
+
+/// Compute a derivation path from a sha256 hash.
+///
+/// Format is a bit peculiar, it's 9 u32's with the top bit as 0 (for unhardened
+/// derivation). We take each u32 in the hash (big endian) and mask off the top bit.
+/// Then we go over the 8 u32s and make a 8 bit u32 from the top bits.
+///
+/// This is because the ChildNumber is a enum u31 where the top bit is used to
+/// indicate hardened or not, so we can't just do the simple thing.
+///
+/// Copied from https://github.com/sapio-lang/sapio/blob/072b8835dcf4ba6f8f00f3a5d9034ef8e021e0a7/ctv_emulators/src/lib.rs
+pub fn hash_to_child_vec(h: sha256::Hash) -> Vec<ChildNumber> {
+    let a: [u8; 32] = h.into_inner();
+    let b: [[u8; 4]; 8] = unsafe { std::mem::transmute(a) };
+    let mut c: Vec<ChildNumber> = b
+        .iter()
+        // Note: We mask off the top bit. This removes 8 bits of entropy from the hash,
+        // but we add it back in later.
+        .map(|x| (u32::from_be_bytes(*x) << 1) >> 1)
+        .map(ChildNumber::from)
+        .collect();
+    // Add a unique 9th path for the MSB's
+    c.push(
+        b.iter()
+            .enumerate()
+            .map(|(i, x)| (u32::from_be_bytes(*x) >> 31) << i)
+            .sum::<u32>()
+            .into(),
+    );
+    c
 }
 
 pub fn concat<T>(mut list: Vec<T>, val: Option<T>) -> Vec<T> {